# resources/public/swagger.yaml
openapi: 3.0.1
info:
  title: Standup bot api
  description: Standup bot api
  contact:
    email: sashaaro@gmail.com
  version: 1.0.0
tags:
  - name: auth
    description: Everything about auth
servers:
  - url: https://standup.botenza.com/api
paths:
  /auth/session:
    get:
      tags:
        - auth
      operationId: getSession
      responses:
        200:
          description: Get session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: Session not found
          content: {}
      x-codegen-request-body-name: body
  /auth/logout:
    get:
      tags:
        - auth
      operationId: logout
      responses:
        301:
          description: Redirect
          content: {}
        200:
          description: Success logout
          content: {}
  /team:
    get:
      tags:
        - team
      operationId: getTeams
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
    post:
      tags:
        - team
      operationId: createTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        400:
          description: Fail
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidationError'
  /team/{id}:
    get:
      tags:
        - team
      operationId: getTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
    put:
      tags:
        - team
      operationId: updateTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        204:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        400:
          description: Fail
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidationError'
  /team/{id}/toggle:
    patch:
      tags:
        - team
      operationId: toggle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
  /user:
    get:
      tags:
        - user
      operationId: getUsers
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /timezone:
    get:
      tags:
        - timezone
      operationId: getTimezones
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Timezone'
  /channel:
    get:
      tags:
        - channel
      operationId: getChannels
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: integer
          format: int64
    Team:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        isEnabled:
          type: boolean
        timezone:
          $ref: '#/components/schemas/Timezone'
        reportChannel:
          $ref: '#/components/schemas/Channel'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
    Question:
      type: object
      properties:
        id:
          type: integer
        index:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        isEnabled:
          type: boolean
        createdAt:
          type: date
    Option:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        updatedAt:
          type: date
        createdAt:
          type: date
    ValidationError:
      type: object
      properties:
        children:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        constraints:
          type: object
          additionalProperties:
            type: string
        property:
          type: string
        #value:
        #  type: any
    Timezone:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: integer
          format: int64
        utc_offset:
          type: object
          properties:
            hours:
              type: integer
            minutes:
              type: integer
    Channel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: integer
          format: int64